Added check to avoid floating point error during reconstruction of parallel runs
PROBLEM:
When the BC is used on parallel, the reconstruction calls initilise() for every processor. When the processor has no eddies in it, smptMax(max(sigma_))=-GREAT and mag(UBulk_)=0. Then the calculation of avgWindow throws floating point exception.
SOLUTION:
The values of sigma_ and UBulk_ are checked before the operation, returning the appropriate value in each case.